requirements.txt
absl-py==0.8.0
aniso8601==8.0.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
attrs==17.4.0
audioread==2.1.8
Automat==0.6.0
bleach==2.1.2
Brlapi==0.6.6
cachetools==4.0.0
certifi==2018.1.18
cffi==1.13.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.1
command-not-found==0.3
constantly==15.1.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
deluge===2.0.3-2-201906121747-ubuntu18.04.1
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.2.3.post1
ffmpeg==1.4
Flask==1.1.1
Flask-Login==0.4.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
gast==0.3.2
GeoIP==1.3.2
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
Google-Images-Search==1.0.1
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.13.2
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
kazam==1.4.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
librosa==0.7.0
llvmlite==0.29.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.3
multiprocessing-logging==0.2.4
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
notebook==5.2.2
numba==0.45.1
numpy==1.18.0
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.2.30
pafy==0.5.3.1
PAM==0.4.2
pandas==0.25.3
pandocfilters==1.4.2
passlib==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.2.1
pipenv==2018.11.26
prompt-toolkit==1.0.15
protobuf==3.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyfiglet==0.8.post1
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
pymongo==3.9.0
PyMySQL==0.9.3
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.2
pyRFC3339==1.0
pyserial==3.4
PySoundFile==0.9.0.post1
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
python-libtorrent==1.1.5
python-resize-image==1.1.19
pytube==9.5.3
pytz==2019.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
rencode==1.0.5
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
resampy==0.2.2
rsa==4.0
scikit-learn==0.21.3
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==16.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.13.2
six==1.13.0
sk-video==1.1.8
SoundFile==0.10.2
sox==1.3.3
SQLAlchemy==1.3.9
system-service==0.3
systemd-python==234
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
terminado==0.7
testpath==0.3.1
tf-estimator-nightly==1.14.0.dev2019060501
tornado==4.5.3
traitlets==4.3.2
Twisted==17.9.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
wadllib==1.3.2
Wave==0.0.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.16.0
wrapt==1.11.2
xkit==0.0.0
youtube-dl==2019.11.28
zope.interface==4.3.2
